@@ -1,7 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
-
-import { of } from 'rxjs/observable/of';
+import { Observable ,  of } from 'rxjs';
 
 export class Interes {
   id: number;
