@@ -3,8 +3,7 @@ import { CloudOptions, CloudData, ZoomOnHoverOptions } from 'angular-tag-cloud-m
 
 @Component({
   selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css'],
+  templateUrl: './app.component.html'
 })
 export class AppComponent {
   options: CloudOptions = {
