@@ -1,3 +1,11 @@
 export const environment = {
-  production: true
+  production: true,
+  firebase: {
+    apiKey: 'AIzaSyBxb7VYNOn9U1xgk0V1sTUA2jYKAEGSC34',
+    authDomain: 'interests-6d04f.firebaseapp.com',
+    databaseURL: 'https://interests-6d04f.firebaseio.com',
+    projectId: 'interests-6d04f',
+    storageBucket: 'interests-6d04f.appspot.com',
+    messagingSenderId: '1002051485129',
+  }
 };
