@@ -4,5 +4,13 @@
 // The list of which env maps to which file can be found in `.angular-cli.json`.
 
 export const environment = {
-  production: false
+  production: false,
+  firebase: {
+    apiKey: 'AIzaSyBxb7VYNOn9U1xgk0V1sTUA2jYKAEGSC34',
+    authDomain: 'interests-6d04f.firebaseapp.com',
+    databaseURL: 'https://interests-6d04f.firebaseio.com',
+    projectId: 'interests-6d04f',
+    storageBucket: 'interests-6d04f.appspot.com',
+    messagingSenderId: '1002051485129',
+  },
 };
